% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{agent_set}
\alias{agent_set}
\title{Create an agent set reporter}
\usage{
agent_set(vars, agents)
}
\arguments{
\item{vars}{A string or vector/list of strings with the variable names of the agent(s).}

\item{agentset}{A string specifying the agent or agentset to be queried.}
}
\description{
Create an agent set reporter to set agent reporters in `nl_experiment` or
  `nl_set_agent_reports`
}
\examples{
experiment <- nl_experiment(
  model_file = file.path(nl_netlogo_path(),
     "models/Sample models/Networks/Preferential attachment.nlogo"),
  max_ticks = 10,
  export_view = TRUE,
  agents_after = list(
    vertices = agent_set(c("who", "xcor", "ycor"), "turtles")
  )
)
}
\keyword{internal}

