% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{criteria}
\alias{criteria}
\title{Criteria expression from a list of expressions}
\usage{
criteria(...)
}
\arguments{
\item{...}{expressions}
}
\description{
Used in evaluate element in \code{\link{nl_experiment}} or
  \code{\link{nl_set_measures}}
}
\details{
Must evaluate to a numeric
}
\keyword{internal}
