% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{measures}
\alias{measures}
\title{Convert measure list to named vector}
\usage{
measures(...)
}
\arguments{
\item{...}{Named character values}
}
\description{
Convert measure list to named vector
}
\keyword{internal}

