% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_default_mapping}
\alias{nl_default_mapping}
\title{Default mapping from R names to NetLogo variables}
\usage{
nl_default_mapping(param_values)
}
\arguments{
\item{param_values}{Parameter values in list or data frame}
}
\value{
Named vector with default mapping.
  Use as function argument in \code{\link{nl_experiment}} mapping.
}
\description{
Creates mapping with simple rule:
changes every character
  \code{_.} to \code{?} and
  \code{_} to \code{-}.
}
\examples{

param_values = list(
  world_size = 50,
  population = 80,
  max_align_turn = c(1, 5, 20),
  max_cohere_turn = c(1, 3, 20),
  max_separate_turn = c(1, 1.5, 20),
  vision = c(1, 3, 10),
  minimum_separation = c(1, 3, 10),
  .dummy = c(1:0)
)

nl_default_mapping(param_values)

# Define experiment mapping with a function instead of named vector:
experiment <- nl_experiment(
  model_file = "models/Sample Models/Biology/Flocking.nlogo",

  param_values = list(
    world_size = 50,
    population = 80,
    max_align_turn = c(1, 5, 20),
    max_cohere_turn = c(1, 3, 20),
    max_separate_turn = c(1, 1.5, 20),
    vision = c(1, 3, 10),
    minimum_separation = c(1, 3, 10),
    .dummy = c(1:0)
  ),
  mapping = nl_default_mapping
)

# check experiment parameter names mapping
cbind(experiment$mapping)

}
