% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlevaluate.R
\name{nl_eval_tracer}
\alias{nl_eval_tracer}
\title{Iterations call-back factory}
\usage{
nl_eval_tracer(verbose = TRUE)
}
\arguments{
\item{verbose}{When TRUE adding new data will print the line}
}
\description{
Iterations call-back factory
}

