% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_experiment}
\alias{nl_experiment}
\title{Create NetLogo experiment object}
\usage{
nl_experiment(model_file, max_ticks = NULL, while_condition = NULL,
  repetitions = 1, step_measures = NULL, run_measures = NULL,
  mapping = NULL, param_values = NULL, export_view = FALSE,
  export_world = FALSE)
}
\arguments{
\item{model_file}{An absolute path to your NetLogo model file (.nlogo)}

\item{max_ticks}{Number of iterations to run.
This parameter is used when while_condition is not defined.}

\item{while_condition}{A string with a NetLogo conditional reporter.
(for example: "ticks < 100")}

\item{repetitions}{How many times to run the model with the same parameters.
It is set to 1 by default. Result data sets will include run_id as
additional variable to identify the specific runs.}

\item{step_measures}{Measures per each simulation step in a named character
vector. Use measures() function to construct measures in right format.}

\item{run_measures}{Measures per each simulation run in a named character
vector. Use measures() function to construct measures in right format.}

\item{mapping}{Mapping between R and NetLogo parameters
in named character vector.
For example: c(diffusion_rate = "diffusion-rate", population = "population")}

\item{param_values}{A data.frame with parameter values or
a list of values to be expanded to all combinations of values}

\item{export_view}{If set to TRUE, the views will be exported to
a png image files for each run (when running the experiment)}

\item{export_world}{If set to TRUE, the world will be exported to
a csv file for each run}
}
\value{
NetLogo experiment object
}
\description{
Create NetLogo experiment object
}

