% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_export_path}
\alias{nl_export_path}
\title{Get and set export path}
\usage{
nl_export_path(export_path = NULL)
}
\arguments{
\item{export_path}{target folder to export files}
}
\description{
Get and set export path
}

