% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlresult.R
\name{nl_get_result}
\alias{nl_get_result}
\title{Get step observations joined with parameter space values}
\usage{
nl_get_result(result, add_parameters = TRUE, type = "run",
  sub_type = NULL)
}
\arguments{
\item{result}{NetLogo result object}

\item{add_parameters}{Add parameter values from parameter space to the results}

\item{type}{Observation type: "run", "step", "criteria", "agents_after", "patches_after"}

\item{sub_type}{Observation sub-type}
}
\description{
Get step observations joined with parameter space values
}

