% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlresult.R
\name{nl_get_run_result}
\alias{nl_get_run_result}
\title{Get run observations joined with parameter space values}
\usage{
nl_get_run_result(result, add_parameters = TRUE)
}
\arguments{
\item{result}{NetLogo result object}

\item{add_parameters}{If parameter values should be appended to the results}
}
\description{
Get run observations joined with parameter space values
}

