% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlimport.R
\name{nl_import_sliders}
\alias{nl_import_sliders}
\title{Import sliders from NetLogo model file}
\usage{
nl_import_sliders(experiment, max_values = 20)
}
\arguments{
\item{experiment}{NetLogo experiment object}

\item{max_values}{Maximum values per parameter}
}
\value{
A list with slider data, suggested parameter sets and mapping
}
\description{
Reads NetLogo model file and parses slider section
}
\details{
Imports parameter names and ranges from sliders defined in
  NetLogo model file.
  Based on information from
    https://github.com/NetLogo/NetLogo/wiki/Model-file-format
    https://github.com/NetLogo/NetLogo/wiki/Widget-Format
}

