% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlrun.R
\name{nl_map_parameter}
\alias{nl_map_parameter}
\title{Internal: maps parameter}
\usage{
nl_map_parameter(experiment, parameter_name)
}
\arguments{
\item{experiment}{Experiment object}

\item{parameter_name}{Parameter name to map}
}
\value{
NetLogo variable name
}
\description{
Internal: maps parameter
}

