% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_netlogo_path}
\alias{nl_netlogo_path}
\title{Get and set netlogo path}
\usage{
nl_netlogo_path(nl_path = NULL)
}
\arguments{
\item{nl_path}{An absolute path to your NetLogo installation
On Windows, for example, something like "C:/Program Files/NetLogo 5.1.0".}
}
\description{
Get and set netlogo path
}

