% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_param_fast}
\alias{nl_param_fast}
\title{Generate a parameter value sets for the FAST method}
\usage{
nl_param_fast(...)
}
\arguments{
\item{...}{Named list of parameter ranges (numeric vectors)}
}
\value{
A data frame with parameter value sets.
}
\description{
Uses \code{\link[fast]{fast_parameters}} from \pkg{fast} package to create
parameter sets for Fourier Amplitute Sensitivity Test (FAST).
}
\details{
Uses only parameters with min != max values to create parameter sets.
Adds dummy variable.
}
\examples{
param_values <- nl_param_fast(
  world_size = 50,
  population = 80,
  max_align_turn = c(1, 5, 20),
  max_cohere_turn = c(1, 3, 20),
  max_separate_turn = c(1, 1.5, 20),
  vision = c(1, 3, 10),
  minimum_separation = c(1, 3, 10)
)
}
\seealso{
Use \code{\link{nl_get_fast_sensitivity}} to get sensitivity data.
  See \link[fast]{fast} package documentation for FAST algorithm details.
  from the simulation results.
  See \code{\link{nl_param_lhs}} for latin hypercube sampling.
}
