% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlimport.R
\name{nl_parse_model}
\alias{nl_parse_model}
\title{Parse NetLogo model file}
\usage{
nl_parse_model(model_file)
}
\arguments{
\item{model_file}{NetLogo model file}
}
\value{
Returns an object of class \code{nl_model}.
  It is a list containing at most the following components:
  \item{ view }{a data frame with NetLogo model view attributes}
  \item{ sliders }{a data frame with NetLogo model sliders attributes}
  \item{ switches }{a data frame with NetLogo model switches attributes}
  \item{ monitors }{a data frame with NetLogo model monitors attributes}
  \item{ plots }{a data frame with NetLogo model plots attributes}
}
\description{
Gets information about widgets (e.g. sliders, monitors, plots) from from NetLogo model file
}
\details{
Imports attributes from sliders defined in
  NetLogo model file.
  Based on information from
    \url{https://github.com/NetLogo/NetLogo/wiki/Model-file-format}
    and
    \url{https://github.com/NetLogo/NetLogo/wiki/Widget-Format}
}

