% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_set_agent_reports}
\alias{nl_set_agent_reports}
\title{Set or change agent reports}
\usage{
nl_set_agent_reports(experiment, agents_before = NULL, agents_after = NULL)
}
\arguments{
\item{experiment}{NetLogo experiment object}

\item{agents_before}{A list of agent reports to be accessed before each run}

\item{agents_after}{A list of agent reports to be accessed after each run}
}
\value{
NetLogo experiment object
}
\description{
Set reporting of variable value(s) of one or more agent(s) as a data.frame
}
\seealso{
To create an experiment object use \code{\link{nl_experiment}}
}

