% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_set_measures}
\alias{nl_set_measures}
\title{Add measures for NetLogo experiment}
\usage{
nl_set_measures(experiment, step = NULL, run = NULL, as.data.frame = TRUE,
  step_transform = NULL)
}
\arguments{
\item{experiment}{NetLogo model object from nl_experiment() function}

\item{step}{NetLogo reporters for each step (reported at every tick)}

\item{run}{NetLogo reporters for each run (reported at end of run)}

\item{as.data.frame}{Reporting in data frame format}

\item{step_transform}{A function to transform data frame result from
step reporters.}
}
\value{
NetLogo experiment object
}
\description{
Add measures for NetLogo experiment
}

