% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_set_param_space}
\alias{nl_set_param_space}
\title{Define parameter space for NetLogo experiment}
\usage{
nl_set_param_space(experiment, param_values = NULL)
}
\arguments{
\item{experiment}{NetLogo experiment object from nl_experiment() function}

\item{param_values}{A data.frame with parameter values or
a list of values to be expanded to all combinations of values}
}
\value{
NetLogo experiment object
}
\description{
Define parameter space for NetLogo experiment
}

