% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_set_run_options}
\alias{nl_set_run_options}
\title{Set run options to a NetLogo experiment}
\usage{
nl_set_run_options(experiment, random_seed = NULL, repetitions = 1,
  max_minutes = 10, setup_commands = "setup", go_command = "go")
}
\arguments{
\item{experiment}{NetLogo experiment object from nl_experiment() function}

\item{random_seed}{Random seed}

\item{repetitions}{Number of repetitions (when random seed is not defined)}

\item{max_minutes}{If max.minutes > 0 the execution stops after the
defined number of minutes (with an error and no return value)}

\item{setup_commands}{NetLogo command strings to execute to setup the model}

\item{go_command}{NetLogo command string to execute the step in the model}
}
\value{
NetLogo experiment object
}
\description{
Set run options to a NetLogo experiment
}

