% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_show_params}
\alias{nl_show_params}
\title{Plots parameters with scatter plots}
\usage{
nl_show_params(experiment, cex = 0.7, col = "#000000CC",
  lower.panel = NULL, ...)
}
\arguments{
\item{experiment}{Experiment object}

\item{cex}{Parameter passed to pairs function}

\item{col}{Parameter passed to pairs function}

\item{lower.panel}{Parameter passed to pairs function}

\item{...}{Parameters passed to pairs function}
}
\description{
Plots parameters with scatter plots
}

