% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlresult.R
\name{nl_show_patches}
\alias{nl_show_patches}
\title{Plot multiple patches result}
\usage{
nl_show_patches(result, x_param, y_param = NULL, fill = "pcolor",
  type = "patches_after", sub_type = NULL)
}
\arguments{
\item{result}{NetLogo experiment result object}

\item{x_param}{row parameter}

\item{y_param}{column parameter}

\item{fill}{variable to control the color (default is pcolor)}

\item{type}{as type from nl_get_result (default is "patches_after)}

\item{sub_type}{as sub_type from nl_get_result (optional - if not the first patches set)}
}
\description{
Plot patches from simualations result
}

