% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlresult.R
\name{nl_show_step}
\alias{nl_show_step}
\title{Plot step measure observations}
\usage{
nl_show_step(result, x = "step_id", y, color = "run_id", x_param = ".",
  y_param = ".", title = NULL, data_filter = NULL, alpha = 1)
}
\arguments{
\item{result}{NetLogo experiment result object}

\item{x}{"step_id" or measure name (as string) to choose for x axis}

\item{y}{measure name as string to plot on y axis}

\item{color}{by default it is based on "run_id" (simulation repetition).
Change to \code{NA} to plot every repetition in black}

\item{x_param}{which parameter to use for faceting horizontally}

\item{y_param}{which parameter to use for faceting vertically}

\item{title}{plot title}

\item{data_filter}{optional subset expression (not quoted) using
parameters, \code{run_id} and \code{step_id}}

\item{alpha}{lines opacity}
}
\description{
Plot observations for each simulation step
}
\seealso{
To get only data and create custom plots see \code{\link{nl_get_result}}
}
