% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlresult.R
\name{nl_show_view}
\alias{nl_show_view}
\title{Show view(s) from a NetLogo result object}
\usage{
nl_show_view(result, param_space_id = NULL, run_id = NULL)
}
\arguments{
\item{result}{NetLogo result object}

\item{param_space_id}{Optional filter on parameter space ID}

\item{run_id}{Optional filter on run ID}
}
\description{
Show view(s) from a NetLogo result object
}

