% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlrun.R
\name{nl_single_export}
\alias{nl_single_export}
\title{Export view and/or world for individual run}
\usage{
nl_single_export(experiment, param_set_id = NA, run_id = 1)
}
\arguments{
\item{experiment}{experiment object}

\item{experiment}{experiment object}
}
\description{
Export view and/or world for individual run
}
\keyword{internal}

