% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\docType{package}
\name{nlexperiment-package}
\alias{nlexperiment}
\alias{nlexperiment-package}
\title{nlexperiment: NetLogo experiments}
\description{
Define and run NetLogo (Wilensky 1999) model experiments.
}
\details{
\pkg{RNetLogo} package (Thiele 2014) opens countless possibilities
by connecting R with NetLogo.
But it does require some R programming to define and run
model experiments.
The purpose of \bold{nlexperiment} is to make exploring NetLogo models
as simple as possible
while keeping complex functionalities available to advanced R users.
User can start with
experiment (analogous to NetLogo Behavior Space),
run the experiment, explore the results
and return to refine the experiment definition.

Functions in \bold{nlexperiment} assume the following steps:
\itemize{
\item Define NetLogo experiment object with parameter sets,
   measures and other related simulation options
   (see \code{\link{nl_experiment}} function).
\item Run experiment (see \code{\link{nl_run}}).
  The result of running an experiment keeps original
  experiment definition
  along with the simulation results and makes the process of model analysis
  more concise and reproducible.
  To run the simulation in parallel working processes
  just use the \code{parallel} attribute in \code{nl_run} function.
\item Analyse and present results of simulation(s).
\item When additional questions pop out, changes to
  experiment will be needed.
  Refine the original definition of the experiment by
  changing only parameter sets (\code{\link{nl_set_param_values}}),
  set different measures (\code{\link{nl_set_measures}}) or set other simulation
  options (\code{\link{nl_set_run_options}}).
}
}
\examples{
\dontrun{
# Set the path to your NetLogo installation
nl_netlogo_path("c:/Program Files (x86)/NetLogo 5.1.0/")

# Create NetLogo experiment of Net Logo Fire model
experiment <- nl_experiment(
  model_file = file.path(nl_netlogo_path(),
      "models/Sample Models/Earth Science/Fire.nlogo")
  while_condition = "any? turtles",
  repetitions = 10,
  run_measures = measures(
    percent_burned = "(burned-trees / initial-trees) * 100",
    progress = "max [pxcor] of patches with [pcolor > 0 and pcolor < 55]"
  ),
  param_values = list(
    density = seq(from = 55, to = 62, by = 1)
  )
)

# Run the experiment using multi-core processing
result <- nl_run(experiment, parallel = TRUE)

# Get observations data frame
dat <- nl_get_run_result(result)

# plot percent burned by density
library(ggplot2)
ggplot(dat, mapping = aes(x = factor(density), y = percent_burned) ) +
  geom_violin() +
  geom_jitter(position = position_jitter(width = 0.2), alpha = 0.3)
}
}
\references{
Wilensky, U. (1999) NetLogo. \url{http://ccl.northwestern.edu/netlogo/}. Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.

  Thiele, J. (2014) R Marries NetLogo: Introduction to the RNetLogo Package. Journal of Statistical Software 58(2) 1-41. \url{http://www.jstatsoft.org/v58/i02/}

  The ideas and principles of NetLogo experiment definition is taken from
  the NetLogo's Behavior Space tool
  \url{http://ccl.northwestern.edu/netlogo/docs/behaviorspace.html}

  The parallel implementation of \code{nl_run} is based on the RNetLogo vignette
  \url{https://cran.r-project.org/web/packages/RNetLogo/vignettes/parallelProcessing.pdf}
}

