% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\docType{package}
\name{nlexperiment}
\alias{nlexperiment}
\alias{nlexperiment-package}
\title{nlexperiment: NetLogo experiments}
\description{
Define and run NetLogo (Wilensky 1999) model experiments.
}
\details{
\link{RNetLogo} package (Thiele 2014) makes it possible to run
NetLogo models and get values from reporters.
But it requires some additional programming to define and run
model experiments (more than using
NetLogo Behavior Space for example).

Functions in \bold{nlexperiment} assume the following steps:
\itemize{
\item Define NetLogo experiment object with parameter space definition,
   selected measures and other related simulation options
   (see \code{\link{nl_experiment}} function).
\item Run experiment (see \code{\link{nl_run}}).
  The result of running an experiment keeps original
  experiment definition
  along with the simulation results and makes the process of model analysis
  more concise and reproducible.
\item Analyse and present results of simulation(s).
\item When additional questions pop out, changes to
  experiment will be needed.
  Refine the original definition of the experiment by
  changing for example only parameter space (\code{\link{nl_set_param_space}}),
  set different measures (\code{\link{nl_set_measures}}) or set other simulation
  options (\code{\link{nl_set_run_options}}).
}
}
\examples{
\dontrun{
# Set the path to your NetLogo installation
nl_netlogo_path("c:/Program Files (x86)/NetLogo 5.1.0/")

# Fire model is included in NetLogo sample models:
fire_model <- file.path(nl_netlogo_path(), "models/Sample Models/Earth Science/Fire.nlogo")

# Create a NetLogo experiment object
experiment <- nl_experiment(
  model_file = fire_model,
  while_condition = "any? turtles",
  repetitions = 10,
  run_measures = measures(
    percent_burned = "(burned-trees / initial-trees) * 100",
    progress = "max [pxcor] of patches with [pcolor > 0 and pcolor < 55]"
  ),
  param_values = list(
    density = seq(from = 55, to = 62, by = 1)
  )
)

# Run the experiment
result <- nl_run(experiment)

dat <- nl_get_run_result(result)

library(ggplot2)
# plot percent burned by density
ggplot(dat, mapping = aes(x = factor(density), y = percent_burned) ) +
  geom_violin() +
  geom_jitter(position = position_jitter(width = .2))
}
}
\references{
Wilensky, U. (1999) NetLogo. \url{http://ccl.northwestern.edu/netlogo/}. Center for Connected Learning and Computer-Based Modeling, Northwestern University. Evanston, IL.

  Thiele, J. (2014) R Marries NetLogo: Introduction to the RNetLogo Package. Journal of Statistical Software 58(2) 1-41. \url{http://www.jstatsoft.org/v58/i02/}

  This package is analogous to the Behavior Space functionality of NetLogo
  \url{http://ccl.northwestern.edu/netlogo/docs/behaviorspace.html}
}

