% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nlexperiment.roadmap}
\alias{nlexperiment.roadmap}
\title{Roadmap}
\description{
Roadmap
}
\details{
\itemize{
\item measure_condition
\item derivative measures
\item Examples: bayes, GA
\item
\item
}
}
\keyword{internal}

