% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{patch_set}
\alias{patch_set}
\title{Create a patch set reporter}
\usage{
patch_set(vars = c("pxcor", "pycor", "pcolor"), patches = "patches")
}
\arguments{
\item{vars}{A string or vector/list of strings with the variable names of the agent(s).}

\item{patches}{A string specifying the patches to be queried.}
}
\description{
Create a patch set reporter to set patch reporters in `nl_experiment` or
  `nl_set_agent_reports`
}
\keyword{internal}

