% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{print.nl_experiment}
\alias{print.nl_experiment}
\title{Print NetLogo experiment object}
\usage{
\method{print}{nl_experiment}(x, ...)
}
\arguments{
\item{x}{NetLogo experiment object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print NetLogo experiment object
}
