% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{measures}
\alias{measures}
\title{Convert measure list to named vector}
\usage{
measures(...)
}
\arguments{
\item{...}{Named character values}
}
\description{
Convert measure list to named vector
}
\examples{
experiment <- nl_experiment(
  model_file = "models/Sample Models/Earth Science/Fire.nlogo",
  while_condition = "any? turtles",
  repetitions = 20,
  run_measures = measures(
    percent_burned = "(burned-trees / initial-trees) * 100",
    progress = "max [pxcor] of patches with [pcolor > 0 and pcolor < 55]"
  ),
  param_values = list(
    density = seq(from = 55, to = 62, by = 1)
  )
)
}
\keyword{internal}
