% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlevaluate.R
\name{nl_eval_init}
\alias{nl_eval_close}
\alias{nl_eval_init}
\alias{nl_eval_run}
\title{Evaluate experiment with specific parameters}
\usage{
nl_eval_init(experiment, parallel = FALSE, max_cores = NULL)

nl_eval_run(param_set, experiment, criteria = NULL, print_progress = FALSE,
  call_back = NULL, parallel = FALSE, cluster = NULL)

nl_eval_close(parallel = FALSE, cluster = NULL)
}
\arguments{
\item{experiment}{NetLogo experiment object (see \code{\link{nl_experiment}})}

\item{parallel}{If TRUE nl_eval_init returns cluster object which should be
passed to nl_eval_run and nl_eval_close.}

\item{max_cores}{If not defined all available cores are used.}

\item{param_set}{parameter set (a list of parameters with values)}

\item{criteria}{Which experiment evaluation criteria to be returned}

\item{print_progress}{print evaluation progress}

\item{call_back}{A call-back function for tracing result in optimization
processes}

\item{cluster}{Required for parallel execution (nl_eval_init returns
cluster object)}
}
\description{
Runs experiment as with \code{\link{nl_run}} but requires started NetLogo instance
with loaded model.

Stop NetLogo instance
}
\details{
Use when parameters depend on previous evaluation (like in
  parameter fitting / callibration / optimization methods).
  It can use the same experiment object as \code{nl_run} function
  where evaluation criteria should be defined
  (see \code{\link{nl_experiment}} or \code{\link{nl_set_measures}}).
}

