% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_experiment}
\alias{nl_experiment}
\title{Create NetLogo experiment object}
\usage{
nl_experiment(model_file, max_ticks = NULL, while_condition = NULL,
  repetitions = 1, random_seed = NULL, step_measures = NULL,
  run_measures = NULL, mapping = NULL, param_values = NULL,
  agents_after = NULL, patches_after = NULL, export_view = FALSE,
  export_world = FALSE, data_handler = NULL)
}
\arguments{
\item{model_file}{An absolute path to your NetLogo model file (.nlogo)}

\item{max_ticks}{Number of iterations to run.
This parameter is used when while_condition is not defined.}

\item{while_condition}{A string with a NetLogo conditional reporter.
(for example: "ticks < 100")}

\item{repetitions}{How many times to run the model with the same parameters.
It is set to 1 by default. Result data sets will include run_id as
additional variable to identify the specific runs. To change repetitions
of existing experiment object use \code{\link{nl_set_run_options}}}

\item{random_seed}{If defined, random seed will be set for each run.
Note: using random seed and repetitions > 1 does not make sense.}

\item{step_measures}{Measures per each simulation step in a named character
vector. Use measures() function to construct measures in right format.
To change step measures
of existing experiment object use \code{\link{nl_set_measures}}}

\item{run_measures}{Measures per each simulation run in a named character
vector. Use measures() function to construct measures in right format.
To change run measures
of existing experiment object use \code{\link{nl_set_measures}}}

\item{mapping}{Mapping between R and NetLogo parameters
in named character vector.
For example: c(diffusion_rate = "diffusion-rate", population = "population")}

\item{param_values}{A data.frame with parameter values or
a list of values to be expanded to all combinations of values}

\item{agents_after}{Agents reporters see \code{\link{nl_set_agent_reports}}}

\item{patches_after}{Patches reporters see \code{\link{nl_set_agent_reports}}}

\item{export_view}{If set to TRUE, the views will be exported to
a png image files for each run (when running the experiment)}

\item{export_world}{If set to TRUE, the world will be exported to
a csv file for each run}

\item{data_handler}{Function to handle observations. If handler is defined
the observations will not be stored in result elements when running
the experiment with `nl_run` function.}
}
\value{
NetLogo experiment object
}
\description{
Use this function to create NetLogo experiment object.
}
\examples{
experiment <- nl_experiment(
  model_file = "my_model.nlogo",
  while_condition = "any? turtles",
  repetitions = 20,
  run_measures = measures(
    percent_burned = "(burned-trees / initial-trees) * 100",
    progress = "max [pxcor] of patches with [pcolor > 0 and pcolor < 55]"
  ),
  param_values = list(
    density = seq(from = 55, to = 62, by = 1)
  )
)
}
\seealso{
To run experiment use \code{\link{nl_run}}.
  To change existing
  experiment object see \code{\link{nl_set_measures}},
  \code{\link{nl_set_run_options}} and
  \code{\link{nl_set_param_space}}.
}

