% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_get_param_range}
\alias{nl_get_param_range}
\title{Get ranges of experiment parameter sets}
\usage{
nl_get_param_range(experiment, as.data.frame = FALSE)
}
\arguments{
\item{experiment}{NetLogo experiment object}

\item{as.data.frame}{Return in a data frame}
}
\value{
A list with lower and upper values for all parameters in
  experiment parameter set.
  When as.data.frame is specified
  a data frame with lower and upper columns.
}
\description{
Upper and lower value for each parameter in experiment parameter sets
}

