% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlresult.R
\name{nl_get_result}
\alias{nl_get_criteria_result}
\alias{nl_get_result}
\alias{nl_get_run_result}
\alias{nl_get_step_result}
\title{Get observations joined with parameter values}
\usage{
nl_get_result(result, add_parameters = TRUE, type = "run",
  sub_type = NULL, ...)

nl_get_run_result(result, add_parameters = TRUE, ...)

nl_get_step_result(result, add_parameters = TRUE, ...)

nl_get_criteria_result(result, add_parameters = TRUE, ...)
}
\arguments{
\item{result}{A nlexperiment result object}

\item{add_parameters}{Add parameter values from parameter space to the results}

\item{type}{Observation type: "run", "step", "criteria", "agents_after", "patches_after"
See \code{\link{nl_run}} for simulations result structure.}

\item{sub_type}{Observation sub-type (in case of individual agents measures
the sub type is a name of the measure)}

\item{...}{expressions to transform resulting data frame}
}
\description{
Observations are stored in result object only with references to
parameter sets (param_set_id). \code{nl_get_result} joins the data
with actual parameters used for each observation.
}

