% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_param_lhs}
\alias{nl_param_lhs}
\title{Create parameter sets with latin hypercube sampling}
\usage{
nl_param_lhs(n, ...)
}
\arguments{
\item{n}{Number of parameter sets}

\item{...}{Named list of parameter ranges (numeric vectors)}
}
\description{
Parameter sets are created with \code{lhs} function from \pkg{tgp} package
}

