% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlrun.R
\name{nl_run}
\alias{nl_run}
\title{Run NetLogo experiment}
\usage{
nl_run(experiment, print_progress = FALSE, gui = FALSE, parallel = FALSE,
  max_cores = NULL)
}
\arguments{
\item{experiment}{NetLogo experiment object}

\item{print_progress}{Set to TRUE if you want to follow the progress in the console}

\item{gui}{Start NetLogo with GUI (by default NetLogo is run in headless mode)}

\item{parallel}{Runs experiment in parallel worker processes
(requires \link[parallel]{parallel} package)}

\item{max_cores}{(optional) only relevant if parallel = TRUE.
If not defined all available processors will be used}
}
\value{
Returns an object of class \code{nl_result}.
  It is a list containing at most the following components:
  \item{ step }{a data frame with observations based on temporal (step) measures.
    It includes at least
    param_set_id (id of parameter set),
    run_id (ID of simulation repetition ),
    step_id (ID of simulation step ),
    and columns named after the temporal measures}
  \item{ run }{a data frame with observations based on final run measures.
    It includes at least
    param_set_id (id of parameter set),
    run_id (ID of simulation repetition ),
    and columns named after the temporal measures}
  \item{ agents_after }{a data frame with observations based on agents
    after each simulation run }
  \item{ agents_before }{a data frame with observations based on agents
    before each simulation run}
  \item{ patches_after }{a data frame with observations based on patches
    after each simulation run }
  \item{ patches_before }{a data frame with observations based on patches
    before each simulation run}
  \item{ criteria }{a data frame with values provided by
  criteria expressions (\code{eval_criteria} in experiment definition
  possibly aggregated by \code{eval_aggregate_fun})  and additional
  criteria defined by \code{eval_mutate} expressions
  }
  \item{ export }{a filename list with reference to
    parameter sets and simulation repetitions}
  \item{ duration }{time spent to complete the experiment (in \code{\link{difftime}})}
  \item{ experiment }{original NetLogo experiment object used}
}
\description{
Runs NetLogo model for defined every parameter and repetitions. Returns
  a list of data frames for each measure defined in experiment.
}
\details{
Model is run for each parameter combination
  defined in parameter sets If \code{repetition} (defined in experiment)
  is greater than \code{1} then
  each run for a parameter set is repeated accordingly.
  Before each run the parameters are set and setup procedure(s) are called.
  After each run criteria function(s) are calculated (if defined)


  Use parallel option if there are more than a few runs per processor core.
}
\seealso{
See \code{\link{nl_experiment}} for creating NetLogo experiment object.
}
