% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlrun.R
\name{nl_run}
\alias{nl_run}
\title{Run NetLogo experiment}
\usage{
nl_run(experiment, print_progress = FALSE, gui = FALSE, parallel = FALSE,
  max_cores = NULL)
}
\arguments{
\item{experiment}{NetLogo experiment object}

\item{print_progress}{Set to TRUE if you want to follow the progress in the console}

\item{gui}{Start NetLogo with GUI (by default NetLogo is run in headless mode)}

\item{parallel}{Runs experiment in parallel worker processes
(requires \link[parallel]{parallel} package)}

\item{max_cores}{(optional) only relevant if parallel = TRUE.
If not defined all available processors will be used}
}
\value{
Returns a list of values which depend on measures defined for each step
  and/or for each run, parameter space, repetitions and export options.
  \itemize{
  \item Element \code{step} is a result of step measures.
  \item Element \code{run} is a result of run measures.
  \item Element \code{export} contains file names of exported views
  and worlds.
  \item Element \code{duration} is a time spent to complete the experiment.
  \item Element \code{experiment} is the NetLogo experiment object used for simulation.
  }
}
\description{
Run NetLogo experiment
}
\details{
Model is run for each parameter combination
  defined in parameter space. If \code{repetition} (defined in experiment)
  is greater than \code{1} then
  each parameter combination is repeated accordingly.

  Use parallel option if there are more than a few runs per processor core.
}
\seealso{
See \code{\link{nl_experiment}} for creating NetLogo experiment object.
}

