% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_run}
\alias{nl_run}
\title{Run NetLogo experiment}
\usage{
nl_run(experiment, gui = FALSE)
}
\arguments{
\item{experiment}{NetLogo experiment object}

\item{gui}{Start NetLogo with GUI (by default NetLogo is run in headless mode)}
}
\value{
A list of data frames with observations (per run and per simulation step)
}
\description{
Run NetLogo experiment
}
\details{
NetLogo is started and the model is run for each tuple of parameters
  defined in parameter space. When repetition (in run options) is > 1 than
  each tuple run is repeated accordingly.
  The result depends on measures defined for each step and for each run.
}

