% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nl_set_measures}
\alias{nl_set_measures}
\title{Set or change measures of existing NetLogo experiment}
\usage{
nl_set_measures(experiment, step = NULL, run = NULL, eval_criteria = NULL,
  eval_aggregate_fun = NULL, eval_mutate = NULL, as.data.frame = TRUE,
  step_transform = NULL)
}
\arguments{
\item{experiment}{NetLogo experiment object}

\item{step}{NetLogo reporters for each step (reported at every tick).
A list of named character vectors. Use \code{\link{measures}} function to get
the correct structure.}

\item{run}{NetLogo reporters for each run (reported at end of run).
A list of named character vectors. Use \code{\link{measures}} function to get
the correct structure.}

\item{eval_criteria}{A criteria calculation expressions.
May use \code{step} or \code{run} data frames to calculate criteria.
Elements from \code{step} should be aggregated.
Must return named numeric vector.}

\item{eval_aggregate_fun}{Aggregate criteria.
It makes sense when when repetitions > 1}

\item{eval_mutate}{Add criteria based on aggregated values}

\item{as.data.frame}{Reporting in data frame format (TRUE by default)}

\item{step_transform}{A function to transform data frame result from
step reporters. When simulation has many steps and only summary
data is needed, step_transform can reduce memory requirements to
run experiment.}
}
\value{
NetLogo experiment object
}
\description{
Set or change measures of existing NetLogo experiment
}
\details{
Values of experiment measures are NetLogo reporters.
  Names of measures will be used in the resulting data frames as
  column names.
}
\seealso{
To create an experiment object use \code{\link{nl_experiment}}
}

