% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlresult.R
\name{nl_show_views_grid}
\alias{nl_show_views_grid}
\title{Show exported views images in grid}
\usage{
nl_show_views_grid(result, x_param = NULL, y_param = NULL)
}
\arguments{
\item{result}{Result from \code{nl_run} function}

\item{x_param}{Name of parameter on x axis}

\item{y_param}{Name of parameter on y axis}
}
\description{
Show exported views images in grid
}

