% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlexperiment.R
\name{nlexperiment.roadmap}
\alias{nlexperiment.roadmap}
\title{Roadmap}
\description{
Roadmap
}
\details{
\itemize{
\item Get agents' data
\item condition changes as a function of parameters?
\item Data handler sample -
\item Parameter space - examples (sampling, calibrating, ...)
}
}
\keyword{internal}

