% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlrun.R
\name{run.steps}
\alias{run.steps}
\title{Run steps (internal doc)}
\description{
\itemize{
\item nl_run | nl_run_parallel
  \itemize{
    \item nl_run_init
    \item nl_single_run
    \itemize{
      \item nl_single_run_setup setting parameters, setup commands
      \item execute NLDoReportWhile | NLDoCommandWhile
      \item get run measures
      \item nl_single_agent_report
      \item nl_single_export
      \item calling external data handler
    }
    \item nl_run_wrap_results
    \item nl_run_end
  }
}
}
\keyword{internal}

